requirements.txt
requests
semanticscholar
pymongo[srv]
click
pyarrow
pandas
xlrd
pymupdf
tqdm
seaborn==0.11.2
pyyaml
scipy
lxml
statannotations
tiktoken
tables
statsmodels
openai
pingouin